Extract custom fields transformed values to a service with cache capabilities
